Loading screen and mneu screnn
