+++ community/keyboard_mouse.py
@@ -15,9 +15,9 @@ class Keyboard_Mouse(controller.Controller):
 		self.game.on_key_release = self.on_key_release
 
 	def on_mouse_press(self, x, y, button, modifiers):
-		if not self.game.mouse_captured:
-			self.game.mouse_captured = True
-			self.game.set_exclusive_mouse(True)
+		if not self.game.window.mouse_captured:
+			self.game.window.mouse_captured = True
+			self.game.window.set_exclusive_mouse(True)
 
 			return
 
@@ -26,7 +26,7 @@ class Keyboard_Mouse(controller.Controller):
 		elif button == pyglet.window.mouse.MIDDLE: self.interact(self.InteractMode.PICK)
 
 	def on_mouse_motion(self, x, y, delta_x, delta_y):
-		if self.game.mouse_captured:
+		if self.game.window.mouse_captured:
 			sensitivity = 0.004
 
 			self.game.player.rotation[0] += delta_x * sensitivity
@@ -38,7 +38,7 @@ class Keyboard_Mouse(controller.Controller):
 		self.on_mouse_motion(x, y, delta_x, delta_y)
 
 	def on_key_press(self, key, modifiers):
-		if not self.game.mouse_captured:
+		if not self.game.window.mouse_captured:
 			return
 
 		if key == pyglet.window.key.D: self.start_move(self.MoveMode.RIGHT)
@@ -61,7 +61,7 @@ class Keyboard_Mouse(controller.Controller):
 		elif key == pyglet.window.key.F10: self.misc(self.MiscMode.TOGGLE_AO)
 
 	def on_key_release(self, key, modifiers):
-		if not self.game.mouse_captured:
+		if not self.game.window.mouse_captured:
 			return
 
 		if key == pyglet.window.key.D: self.end_move(self.MoveMode.RIGHT)
