+++ community/logs/1644067210.7173743.logs
@@ -1,0 +1,14 @@
+[LOGS]
+[2022-02-05 13:20:16,213] [MainProcess/MainThread/INFO] (gamemain.py/__init__) System Info: Python: CPython 3.10.2
+System: AMD64 Windows 10 10.0.19044
+CPU: Intel64 Family 6 Model 165 Stepping 3, GenuineIntel
+Display: Intel(R) UHD Graphics 630 
+4.6.0 - Build 26.20.100.7926
+[2022-02-05 13:20:16,214] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Compiling Shaders
+[2022-02-05 13:20:16,224] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Creating Texture Array
+[2022-02-05 13:20:16,228] [MainProcess/MainThread/INFO] (world.py/__init__) Loading block models
+[2022-02-05 13:20:16,620] [MainProcess/MainThread/INFO] (save.py/load) Loading world
+[2022-02-05 13:20:22,515] [MainProcess/MainThread/INFO] (world.py/__init__) Lighting chunks
+[2022-02-05 13:20:36,289] [MainProcess/MainThread/INFO] (world.py/__init__) Generating chunks
+[2022-02-05 13:20:36,305] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Setting up player & camera
+[2022-02-05 13:20:36,352] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Loading audio
