+++ community/logs/1644067266.2068014.logs
@@ -1,0 +1,28 @@
+[LOGS]
+[2022-02-05 13:21:13,430] [MainProcess/MainThread/INFO] (gamemain.py/__init__) System Info: Python: CPython 3.10.2
+System: AMD64 Windows 10 10.0.19044
+CPU: Intel64 Family 6 Model 165 Stepping 3, GenuineIntel
+Display: Intel(R) UHD Graphics 630 
+4.6.0 - Build 26.20.100.7926
+[2022-02-05 13:21:13,430] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Compiling Shaders
+[2022-02-05 13:21:13,445] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Creating Texture Array
+[2022-02-05 13:21:13,445] [MainProcess/MainThread/INFO] (world.py/__init__) Loading block models
+[2022-02-05 13:21:13,820] [MainProcess/MainThread/INFO] (save.py/load) Loading world
+[2022-02-05 13:21:19,653] [MainProcess/MainThread/INFO] (world.py/__init__) Lighting chunks
+[2022-02-05 13:21:33,490] [MainProcess/MainThread/INFO] (world.py/__init__) Generating chunks
+[2022-02-05 13:21:33,506] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Setting up player & camera
+[2022-02-05 13:21:33,553] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Loading audio
+[2022-02-05 13:21:42,207] [MainProcess/MainThread/INFO] (gamemain.py/__init__) System Info: Python: CPython 3.10.2
+System: AMD64 Windows 10 10.0.19044
+CPU: Intel64 Family 6 Model 165 Stepping 3, GenuineIntel
+Display: Intel(R) UHD Graphics 630 
+4.6.0 - Build 26.20.100.7926
+[2022-02-05 13:21:42,207] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Compiling Shaders
+[2022-02-05 13:21:42,223] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Creating Texture Array
+[2022-02-05 13:21:42,223] [MainProcess/MainThread/INFO] (world.py/__init__) Loading block models
+[2022-02-05 13:21:42,624] [MainProcess/MainThread/INFO] (save.py/load) Loading world
+[2022-02-05 13:21:48,480] [MainProcess/MainThread/INFO] (world.py/__init__) Lighting chunks
+[2022-02-05 13:22:02,479] [MainProcess/MainThread/INFO] (world.py/__init__) Generating chunks
+[2022-02-05 13:22:02,495] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Setting up player & camera
+[2022-02-05 13:22:02,542] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Loading audio
+[2022-02-05 13:22:02,554] [MainProcess/MainThread/INFO] (gamemain.py/on_close) Deleting media player
