+++ community/logs/1644067540.0752838.logs
@@ -1,0 +1,27 @@
+[LOGS]
+[2022-02-05 13:25:46,861] [MainProcess/MainThread/INFO] (gamemain.py/__init__) System Info: Python: CPython 3.10.2
+System: AMD64 Windows 10 10.0.19044
+CPU: Intel64 Family 6 Model 165 Stepping 3, GenuineIntel
+Display: Intel(R) UHD Graphics 630 
+4.6.0 - Build 26.20.100.7926
+[2022-02-05 13:25:46,862] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Compiling Shaders
+[2022-02-05 13:25:46,866] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Creating Texture Array
+[2022-02-05 13:25:46,866] [MainProcess/MainThread/INFO] (world.py/__init__) Loading block models
+[2022-02-05 13:25:47,257] [MainProcess/MainThread/INFO] (save.py/load) Loading world
+[2022-02-05 13:25:53,164] [MainProcess/MainThread/INFO] (world.py/__init__) Lighting chunks
+[2022-02-05 13:26:07,052] [MainProcess/MainThread/INFO] (world.py/__init__) Generating chunks
+[2022-02-05 13:26:07,052] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Setting up player & camera
+[2022-02-05 13:26:07,115] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Loading audio
+[2022-02-05 13:26:25,904] [MainProcess/MainThread/INFO] (gamemain.py/__init__) System Info: Python: CPython 3.10.2
+System: AMD64 Windows 10 10.0.19044
+CPU: Intel64 Family 6 Model 165 Stepping 3, GenuineIntel
+Display: Intel(R) UHD Graphics 630 
+4.6.0 - Build 26.20.100.7926
+[2022-02-05 13:26:25,904] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Compiling Shaders
+[2022-02-05 13:26:25,904] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Creating Texture Array
+[2022-02-05 13:26:25,920] [MainProcess/MainThread/INFO] (world.py/__init__) Loading block models
+[2022-02-05 13:26:26,295] [MainProcess/MainThread/INFO] (save.py/load) Loading world
+[2022-02-05 13:26:32,181] [MainProcess/MainThread/INFO] (world.py/__init__) Lighting chunks
+[2022-02-05 13:26:46,761] [MainProcess/MainThread/INFO] (world.py/__init__) Generating chunks
+[2022-02-05 13:26:46,774] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Setting up player & camera
+[2022-02-05 13:26:46,819] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Loading audio
