+++ community/logs/1644069120.6182787.logs
@@ -1,0 +1,13 @@
+[LOGS]
+[2022-02-05 13:56:40,238] [MainProcess/MainThread/INFO] (gamemain.py/__init__) System Info: Python: CPython 3.10.2
+System: AMD64 Windows 10 10.0.19044
+CPU: Intel64 Family 6 Model 165 Stepping 3, GenuineIntel
+Display: Intel(R) UHD Graphics 630 
+4.6.0 - Build 26.20.100.7926
+[2022-02-05 13:56:40,238] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Compiling Shaders
+[2022-02-05 13:56:40,238] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Creating Texture Array
+[2022-02-05 13:56:40,253] [MainProcess/MainThread/INFO] (world.py/__init__) Loading block models
+[2022-02-05 13:56:40,644] [MainProcess/MainThread/INFO] (save.py/load) Loading world
+[2022-02-05 13:56:46,625] [MainProcess/MainThread/INFO] (world.py/__init__) Lighting chunks
+[2022-02-05 13:57:00,591] [MainProcess/MainThread/INFO] (world.py/__init__) Generating chunks
+[2022-02-05 13:57:00,606] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Setting up player & camera
