+++ community/logs/1644142309.063243.logs
@@ -1,0 +1,28 @@
+[LOGS]
+[2022-02-06 10:11:58,009] [MainProcess/MainThread/INFO] (gamemain.py/__init__) System Info: Python: CPython 3.10.2
+System: AMD64 Windows 10 10.0.19044
+CPU: Intel64 Family 6 Model 165 Stepping 3, GenuineIntel
+Display: Intel(R) UHD Graphics 630 
+4.6.0 - Build 26.20.100.7926
+[2022-02-06 10:11:58,009] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Compiling Shaders
+[2022-02-06 10:11:58,025] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Creating Texture Array
+[2022-02-06 10:11:58,025] [MainProcess/MainThread/INFO] (world.py/__init__) Loading block models
+[2022-02-06 10:11:58,603] [MainProcess/MainThread/INFO] (save.py/load) Loading world
+[2022-02-06 10:12:02,200] [MainProcess/MainThread/INFO] (gamemain.py/__init__) System Info: Python: CPython 3.10.2
+System: AMD64 Windows 10 10.0.19044
+CPU: Intel64 Family 6 Model 165 Stepping 3, GenuineIntel
+Display: Intel(R) UHD Graphics 630 
+4.6.0 - Build 26.20.100.7926
+[2022-02-06 10:12:02,201] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Compiling Shaders
+[2022-02-06 10:12:02,210] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Creating Texture Array
+[2022-02-06 10:12:02,215] [MainProcess/MainThread/INFO] (world.py/__init__) Loading block models
+[2022-02-06 10:12:02,599] [MainProcess/MainThread/INFO] (save.py/load) Loading world
+[2022-02-06 10:12:08,338] [MainProcess/MainThread/INFO] (gamemain.py/__init__) System Info: Python: CPython 3.10.2
+System: AMD64 Windows 10 10.0.19044
+CPU: Intel64 Family 6 Model 165 Stepping 3, GenuineIntel
+Display: Intel(R) UHD Graphics 630 
+4.6.0 - Build 26.20.100.7926
+[2022-02-06 10:12:08,338] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Compiling Shaders
+[2022-02-06 10:12:08,349] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Creating Texture Array
+[2022-02-06 10:12:08,353] [MainProcess/MainThread/INFO] (world.py/__init__) Loading block models
+[2022-02-06 10:12:08,710] [MainProcess/MainThread/INFO] (save.py/load) Loading world
