+++ community/logs/1644142729.1588702.logs
@@ -1,0 +1,15 @@
+[LOGS]
+[2022-02-06 10:18:55,212] [MainProcess/MainThread/INFO] (gamemain.py/__init__) System Info: Python: CPython 3.10.2
+System: AMD64 Windows 10 10.0.19044
+CPU: Intel64 Family 6 Model 165 Stepping 3, GenuineIntel
+Display: Intel(R) UHD Graphics 630 
+4.6.0 - Build 26.20.100.7926
+[2022-02-06 10:18:55,212] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Compiling Shaders
+[2022-02-06 10:18:55,228] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Creating Texture Array
+[2022-02-06 10:18:55,228] [MainProcess/MainThread/INFO] (world.py/__init__) Loading block models
+[2022-02-06 10:18:55,618] [MainProcess/MainThread/INFO] (save.py/load) Loading world
+[2022-02-06 10:19:01,541] [MainProcess/MainThread/INFO] (world.py/__init__) Lighting chunks
+[2022-02-06 10:19:15,372] [MainProcess/MainThread/INFO] (world.py/__init__) Generating chunks
+[2022-02-06 10:19:15,388] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Setting up player & camera
+[2022-02-06 10:19:15,435] [MainProcess/MainThread/INFO] (gamemain.py/__init__) Loading audio
+[2022-02-06 10:20:14,176] [MainProcess/MainThread/INFO] (gamemain.py/on_close) Deleting media player
